Deleted unused files and moved service implementations to corresponding package
